requirements.txt
email_listener==0.1.2
kucoin-futures-python==1.0.6 